% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{addStep,PipR-method}
\alias{addStep,PipR-method}
\alias{addStep}
\title{addStep-methods}
\usage{
\S4method{addStep}{PipR}(object, name, cmd, args = NULL,
  parents = character())
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}

\item{name}{A name for the analysis step. will be used as an identifier}

\item{cmd}{Command(s)/Tool(s) pertaining to the analysis step as character vector. 
If the vector is of length 1 (single string), the command will be recycled
for all \code{args} elements}

\item{args}{Command/Tool arguments pertaining to the analysis step as list of
character vectors. There should be one list element for each command.}

\item{parents}{Character vector specifying the names of other analysis steps the
new step depends on. These steps must already be in the pipeline object.}
}
\value{
the modified pipeline object containing the added analysis step
}
\description{
Add an analysis step to the pipeline
}
\details{
The new analysis step will be registered and scheduled for running using
the \code{\link{run,PipR-method}} command.
}
\author{
Fabian Mueller
}
