% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{changePipeDir,PipR-method}
\alias{changePipeDir,PipR-method}
\alias{changePipeDir}
\title{changePipeDir-methods}
\usage{
\S4method{changePipeDir}{PipR}(object, dir)
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}

\item{dir}{the new pipeline directory}
}
\value{
the modified object with changed directories in the object slots
}
\description{
Change the pipeline directory and it's subdirectories.
}
\details{
Will result in an error if the new pipeline directory is already existing.
}
\author{
Fabian Mueller
}
