% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{checkErrorsInLogFiles}
\alias{checkErrorsInLogFiles}
\title{checkErrorsInLogFiles}
\usage{
checkErrorsInLogFiles(logFns, errorKeys = c("error", "fail", "except"))
}
\arguments{
\item{logFns}{(vector of) log/text file names}
}
\value{
...
}
\description{
Check and report lines containing error keywords in text files
}
\author{
Fabian Mueller
}
