% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{cleanup,PipR-method}
\alias{cleanup}
\alias{cleanup,PipR-method}
\title{cleanup-methods}
\usage{
\S4method{cleanup}{PipR}(object)
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}
}
\value{
the unmodified object (invisible)
}
\description{
Remove temporary and uninformative files.
}
\details{
Everything in the pipelines temporary will be deleted. Additionally, empty files
from the log directoy will be deleted.
}
\author{
Fabian Mueller
}

