% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandR-class.R
\docType{class}
\name{CommandR-class}
\alias{CommandR-class}
\title{CommandR}
\description{
A virtual class for submitting command line calls
}
\details{
For a concrete child class employing system calls see \code{\linkS4class{CommandRsystem}}.
For a concrete child class employing the sun grid engine calls see \code{\linkS4class{CommandRsge}}.
If you want to implement your own child class be sure to at least implement the following functions:
\code{\link{exec,CommandR-method}}.
}
\section{Slots}{

\describe{
  \item{\code{logDir}}{A directory where log and error files are written to}
  \item{\code{doRun}}{Flag indicating whether the analysis should actually run.}
  \item{\code{handlesJobDeps}}{Flag indicating whether the class object can
		actually handle job dependencies on its own
  }
}
}

\section{Methods}{

\describe{
  \item{\code{\link{getLogDir,CommandR-method}}}{Get the logging directory.}
  \item{\code{\link{exec,CommandR-method}}}{Execute a command/job.}
  \item{\code{\link{lexec,CommandR-method}}}{Execute an array of commands/jobs.}
}
}

\author{
Fabian Mueller
}
