% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandRsge-class.R
\docType{class}
\name{CommandRsge-class}
\alias{CommandRsge}
\alias{CommandRsge-class}
\alias{intialize,CommandRsge-method}
\title{CommandRsge Class}
\usage{
CommandRsge(logDir = NULL, req = NULL)
}
\arguments{
\item{logDir}{logging directory}

\item{req}{default resource requirements. Will be specified for each submitted job, if no
other requirements are specified.}
}
\description{
A child class of \code{\linkS4class{CommandR}} implementing job submission using system calls.
}
\details{
Follow this template if you want to create your own \code{\linkS4class{CommandR}} class.
}
\section{Slots}{

Includes all the slots of the \code{\linkS4class{CommandR}} class. Additionally includes:
\describe{
  \item{\code{req}}{Default resource requirements for submitting jobs to the grid engine.}
}
}

\section{Methods}{

implements all the methods of the \code{\linkS4class{CommandR}} class.
}
\author{
Fabian Mueller
}

