% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandRslurm-class.R
\docType{class}
\name{CommandRslurm-class}
\alias{CommandRslurm-class}
\alias{CommandRslurm}
\alias{intialize,CommandRslurm-method}
\title{CommandRslurm Class}
\usage{
CommandRslurm(logDir = NULL, req = NULL)
}
\arguments{
\item{logDir}{logging directory}

\item{req}{default resource requirements. Will be specified for each submitted job, if no
other requirements are specified. Must be a named character vector.}
}
\description{
A child class of \code{\linkS4class{CommandR}} implementing job submission using slurm job scheduling.
}
\details{
Follow this template if you want to create your own \code{\linkS4class{CommandR}} class.
}
\section{Slots}{

Includes all the slots of the \code{\linkS4class{CommandR}} class. Additionally includes:
\describe{
  \item{\code{req}}{Default resource requirements for submitting jobs to the slurm engine.}
}
}

\section{Methods}{

implements all the methods of the \code{\linkS4class{CommandR}} class.
}

\author{
Fabian Mueller
}
