% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandRsystem-class.R
\docType{class}
\name{CommandRsystem-class}
\alias{CommandRsystem-class}
\alias{CommandRsystem}
\alias{intialize,CommandRsystem-method}
\title{CommandRsystem Class}
\usage{
CommandRsystem(logDir = NULL)
}
\arguments{
\item{logDir}{logging directory}
}
\description{
A child class of \code{\linkS4class{CommandR}} implementing job submission using system calls.
}
\details{
Follow this template if you want to create your own \code{\linkS4class{CommandR}} class.
}
\section{Slots}{

Includes all the slots of the \code{\linkS4class{CommandR}} class
}

\section{Methods}{

implements all the methods of the \code{\linkS4class{CommandR}} class.
}

\author{
Fabian Mueller
}
