% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandR-class.R
\docType{methods}
\name{exec,CommandR-method}
\alias{exec,CommandR-method}
\alias{exec}
\title{exec-methods}
\usage{
\S4method{exec}{CommandR}(object, job, result = FALSE, wait = TRUE)
}
\arguments{
\item{object}{\code{\linkS4class{CommandR}} object}

\item{job}{\code{\linkS4class{Job}} object}

\item{result}{Flag indicating whether the result should be returned}

\item{wait}{Flag indicating whether the session should wait for the job to finish before continuing. Will be set to \code{TRUE}
if \code{result} is \code{TRUE}}
}
\value{
\code{\linkS4class{JobResult}} object
}
\description{
Executes a job given the job specifics in the job
}
\details{
For a concrete child class implementations see \code{\link{exec,CommandRsystem-method}} and \code{\link{exec,CommandRsge-method}}
}
\author{
Fabian Mueller
}
