% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandRsystem-class.R
\docType{methods}
\name{exec,CommandRsystem-method}
\alias{exec,CommandRsystem-method}
\title{exec-methods}
\usage{
\S4method{exec}{CommandRsystem}(object, job, result = FALSE, wait = TRUE,
  ...)
}
\arguments{
\item{object}{\code{\linkS4class{CommandRsystem}} object}

\item{job}{\code{\linkS4class{Job}} object}

\item{result}{Flag indicating whether the result should be returned}

\item{wait}{Flag indicating whether the session should wait for the job to finish before continuing. Will be set to \code{TRUE}
if \code{result} is \code{TRUE}}
}
\value{
\code{\linkS4class{JobResult}} object
}
\description{
Executes a job given the job specifics using system calls
}
\details{
Instantiates the method \code{\link{exec,CommandR-method}} for \code{\linkS4class{CommandRsystem}} objects.
}
\author{
Fabian Mueller
}

