% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-class.R
\docType{methods}
\name{getArgs,Job-method}
\alias{getArgs,Job-method}
\alias{getArgs}
\title{getArgs-methods}
\usage{
\S4method{getArgs}{Job}(object)
}
\arguments{
\item{object}{\code{\linkS4class{Job}} object}
}
\value{
Character vector specifying additional arguments to the tool/command
}
\description{
Return the additional arguments for the command call of the object
}
\author{
Fabian Mueller
}
