% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-class.R
\docType{methods}
\name{getCallString,Job-method}
\alias{getCallString}
\alias{getCallString,Job-method}
\title{getCallString-methods}
\usage{
\S4method{getCallString}{Job}(object)
}
\arguments{
\item{object}{\code{\linkS4class{Job}} object}
}
\value{
Character string concatenating the command and arguments
}
\description{
Concatenate the command and arguments into a single string for system calls.
}
\author{
Fabian Mueller
}

