% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-class.R
\docType{methods}
\name{getCmd,Job-method}
\alias{getCmd,Job-method}
\alias{getCmd}
\title{getCmd-methods}
\usage{
\S4method{getCmd}{Job}(object)
}
\arguments{
\item{object}{\code{\linkS4class{Job}} object}
}
\value{
Character specifying the command/tool.
}
\description{
Return the command/tool of the object
}
\author{
Fabian Mueller
}
