% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobResult-class.R
\docType{methods}
\name{getCommand,JobResult-method}
\alias{getCommand,JobResult-method}
\alias{getCommand}
\title{getCommand-methods}
\usage{
\S4method{getCommand}{JobResult}(object)
}
\arguments{
\item{object}{\code{\linkS4class{JobResult}} object}
}
\value{
Command used to run the job as character
}
\description{
Return the command used to run the job.
}
\author{
Fabian Mueller
}
