% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-class.R
\docType{methods}
\name{getDepJobs,Job-method}
\alias{getDepJobs,Job-method}
\alias{getDepJobs}
\title{getDepJobs-methods}
\usage{
\S4method{getDepJobs}{Job}(object)
}
\arguments{
\item{object}{\code{\linkS4class{Job}} object}
}
\value{
The list of other \code{Job} objects that the current job depends on.
}
\description{
Return other jobs the current job depends on
}
\author{
Fabian Mueller
}
