% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{getDir,PipR-method}
\alias{getDir}
\alias{getDir,PipR-method}
\title{getDir-methods}
\usage{
\S4method{getDir}{PipR}(object, type = "base")
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}

\item{type}{Optionally specify the subdirectory. Possible values are
\code{"base", "log", "status", "result", "temp"}}
}
\value{
Character string specifying the path of the directory.
}
\description{
Return the pipeline base directory or one of it's subdirectories
}
\author{
Fabian Mueller
}

