% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobResult-class.R
\docType{methods}
\name{getErr,JobResult-method}
\alias{getErr,JobResult-method}
\alias{getErr}
\title{getErr-methods}
\usage{
\S4method{getErr}{JobResult}(object)
}
\arguments{
\item{object}{\code{\linkS4class{JobResult}} object}
}
\value{
Character vector specifying the errors.
}
\description{
Return the errors of the job.
}
\author{
Fabian Mueller
}
