% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{getGraph,PipR-method}
\alias{getGraph}
\alias{getGraph,PipR-method}
\title{getGraph-methods}
\usage{
\S4method{getGraph}{PipR}(object)
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}
}
\value{
An \code{igraph} object containing the graph
}
\description{
Get a graph representation storing all steps in the pipeline as nodes
and step dependencies as edges.
}
\author{
Fabian Mueller
}
\seealso{
\code{\link{plotGraph,PipR-method}} for plotting the graph
}

