% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-class.R
\docType{methods}
\name{getId,Job-method}
\alias{getId}
\alias{getId,Job-method}
\title{getId-methods}
\usage{
\S4method{getId}{Job}(object)
}
\arguments{
\item{object}{\code{\linkS4class{Job}} object}
}
\value{
Character specifying the id.
}
\description{
Return the id/name of the object
}
\author{
Fabian Mueller
}

