% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandR-class.R
\docType{methods}
\name{getLogDir,CommandR-method}
\alias{getLogDir,CommandR-method}
\alias{getLogDir}
\title{getLogDir-methods}
\usage{
\S4method{getLogDir}{CommandR}(object)
}
\arguments{
\item{object}{\code{\linkS4class{CommandR}} object}
}
\value{
Character specifying the logging directory of the object.
}
\description{
Return the logging directory of an object
}
\author{
Fabian Mueller
}
