% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobResult-class.R
\docType{methods}
\name{getOut,JobResult-method}
\alias{getOut,JobResult-method}
\alias{getOut}
\title{getOut-methods}
\usage{
\S4method{getOut}{JobResult}(object)
}
\arguments{
\item{object}{\code{\linkS4class{JobResult}} object}
}
\value{
Character vector specifying the output.
}
\description{
Return the output of the job.
}
\author{
Fabian Mueller
}
