% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobResult-class.R
\docType{methods}
\name{getStatus,JobResult-method}
\alias{getStatus}
\alias{getStatus,JobResult-method}
\title{getStatus-methods}
\usage{
\S4method{getStatus}{JobResult}(object)
}
\arguments{
\item{object}{\code{\linkS4class{JobResult}} object}
}
\value{
Integer specifying the job's status. \code{0} for success.
}
\description{
Return the status of the job.
}
\author{
Fabian Mueller
}

