% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{getSteps,PipR-method}
\alias{getSteps}
\alias{getSteps,PipR-method}
\title{getSteps-methods}
\usage{
\S4method{getSteps}{PipR}(object, status = "all", graphOrder = FALSE)
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}

\item{status}{Optionally only include complete or incomplete steps in the output.
Possible values are \code{"all", "complete", "incomplete"}.}

\item{graphOrder}{Logical indicating whether the steps should be returned in
graph order, i.e. in the order of a DFS in the pipeline graph.}
}
\value{
Character vector of pipeline step names.
}
\description{
Return the names of pipeline steps
}
\author{
Fabian Mueller
}

