% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandR-class.R
\docType{methods}
\name{initialize,CommandR-method}
\alias{initialize,CommandR-method}
\title{initialize.CommandR}
\usage{
\S4method{initialize}{CommandR}(.Object)
}
\arguments{
\item{.Object}{New instance of \code{CommandR}.}
}
\description{
Initialize a CommandR object
}
\author{
Fabian Mueller
}

