% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{initializePipeDir,PipR-method}
\alias{initializePipeDir}
\alias{initializePipeDir,PipR-method}
\title{initializePipeDir-methods}
\usage{
\S4method{initializePipeDir}{PipR}(object)
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}
}
\value{
the modified object with changed directories in the object slots
}
\description{
Create the pipeline directory and it's subdirectories.
}
\details{
Will result in an error if the pipeline directory is already existing.
}
\author{
Fabian Mueller
}

