% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Job-class.R
\docType{class}
\name{Job-class}
\alias{Job}
\alias{Job-class}
\alias{intialize,Job-method}
\title{Job}
\usage{
Job(cmd, args = character(), id = NULL, dependsOn = list())
}
\arguments{
\item{cmd}{The command or tool to be executed. Must be a single element
character string containing no white spaces.}

\item{args}{A character vector specifying additional arguments to the 
command or tool}

\item{id}{A name or id for the job}

\item{dependsOn}{A list of other \code{Job} objects that the current job depends on.}
}
\description{
A class for storing information on pipeline jobs.
}
\section{Slots}{

\describe{
  \item{\code{cmd}}{
	The command or tool to be executed. Must be a single element character
     string containing no white spaces.
  }
  \item{\code{args}}{
	A character vector specifying additional arguments to the command or tool
  }
  \item{\code{id}}{
	A name or id for the job
  }
  \item{\code{dependsOn}}{
	A list of other \code{Job} objects that the current job depends on.
  }
}
}

\section{Methods}{

\describe{
   \item{\code{
		\link{getId,Job-method},
		\link{getCmd,Job-method},
		\link{getArgs,Job-method},
		\link{getDepJobs,Job-method}
   }}{
     Getters for the \code{id, cmd, args, dependsOn} slots
   }
   \item{\code{\link{getCallString,Job-method}}}{
     concatenate the command and the arguments into one string 
   }
}
}
\author{
Fabian Mueller
}

