% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CommandRsge-class.R
\docType{methods}
\name{lexec,CommandRsge-method}
\alias{lexec,CommandRsge-method}
\title{lexec-methods}
\usage{
\S4method{lexec}{CommandRsge}(object, jobList, result = FALSE, wait = TRUE,
  req = NULL)
}
\arguments{
\item{object}{\code{\linkS4class{CommandR}} object}

\item{jobList}{a list of \code{\linkS4class{Job}} objects}

\item{result}{Flag indicating whether the result should be returned}

\item{wait}{Flag indicating whether the session should wait for the job to finish before continuing. Will be set to \code{TRUE}
if \code{result} is \code{TRUE}}

\item{req}{Resource requirements to the grid engine when submitting the job}
}
\value{
a list of \code{\linkS4class{JobResult}} objects
}
\description{
executes an array of jobs (list execute --> lexec) given the job specifics in the input list of jobs
}
\details{
Instantiates the method \code{\link{lexec,CommandR-method}} for \code{\linkS4class{CommandRsge}} objects.
}
\author{
Fabian Mueller
}

