% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{plotGraph,PipR-method}
\alias{plotGraph,PipR-method}
\alias{plotGraph}
\title{plotGraph-methods}
\usage{
\S4method{plotGraph}{PipR}(object, activeSteps = character())
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}

\item{activeSteps}{Names of active steps to be highlighted in the plot.}
}
\value{
the result of \code{plot}
}
\description{
Plot the pipeline graph representation.
}
\details{
Registered steps will appear in yellow, completed steps in green and active steps in orange.
}
\seealso{
\code{\link{getGraph,PipR-method}} for getting the graph as \code{igraph} object.
}
\author{
Fabian Mueller
}
