% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{resetStep,PipR-method}
\alias{resetStep}
\alias{resetStep,PipR-method}
\title{resetStep-methods}
\usage{
\S4method{resetStep}{PipR}(object, name, delete = TRUE,
  resetDependendSteps = TRUE)
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}

\item{name}{Name for the analysis step to be reset/undone.}

\item{delete}{Logical Indicating whether the step's result directory should be deleted}

\item{resetDependendSteps}{Logical indicating whether subsequent steps that depend on the
reset step should also be reset.}
}
\value{
the modified pipeline object with the reset anlysis step(s)
}
\description{
Reset/undo an analysis step and all of it's dependent steps.
}
\details{
The analysis step and potentially all it's dependent steps will be marked as
'reset' in their status and associated result directories will be deleted.
}
\author{
Fabian Mueller
}

