% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PipR-class.R
\docType{methods}
\name{run,PipR-method}
\alias{run,PipR-method}
\alias{run}
\title{run-methods}
\usage{
\S4method{run}{PipR}(object, cmdr = CommandRsystem(logDir = getDir(object,
  "log")), createDirs = TRUE, saveStatus = TRUE, logCommands = FALSE, ...)
}
\arguments{
\item{object}{\code{\linkS4class{PipR}} object}

\item{cmdr}{\code{\linkS4class{CommandR}} object that will execute all jobs
pertaining to each pipeline step.}

\item{createDirs}{Logical Indicating whether subdirectories for each pipeline step should
be created.}

\item{saveStatus}{Logical indicating whether the \code{\linkS4class{PipR}} object should
save itself and the current pipeline graph with annotated steps
in the pipeline's status directory.}

\item{logCommands}{Logical indicating whether the commands that have been used should be
written to a log file.}

\item{...}{arguments passed on to \code{lexec}}
}
\value{
the modified pipeline object with updated step states.
}
\description{
Run all incomplete analysis steps in the pipeline
}
\details{
The method will run all incomplete analysis steps given a \code{\linkS4class{CommandR}}
object for executing the jobs pertaining to each step
}
\author{
Fabian Mueller
}
