% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/JobResult-class.R
\docType{class}
\name{JobResult-class}
\alias{JobResult}
\alias{JobResult-class}
\alias{intialize,JobResult-method}
\title{JobResult}
\usage{
JobResult(out = character(), err = character(), status = 0L)
}
\arguments{
\item{out}{Character vector storing the output of the job.
character string containing no white spaces.}

\item{err}{Character vector storing errors that occurred during the execution of the job
command or tool}

\item{status}{Integer specifying the job's status. \code{0} for success.}
}
\description{
A class for storing information on the results of pipeline jobs.
}
\section{Slots}{

\describe{
  \item{\code{out}}{
	Character vector storing the output of the job.
  }
  \item{\code{err}}{
	Character vector storing errors that occurred during the execution of the job
  }
  \item{\code{status}}{
	Integer specifying the job's status. \code{0} for success.
  }
}
}

\section{Methods}{

\describe{
   \item{\code{
		\link{getOut,JobResult-method},
		\link{getErr,JobResult-method},
		\link{getStatus,JobResult-method}
   }}{
     Getters for the \code{out, err, status} slots
   }
}
}
\author{
Fabian Mueller
}

